+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### Added
+- Added a `WaitOn` method to `x/erltest/testcase` that will add a dependency
+whose `Wait()` method will be called in parallel with all other "Waitables" when
+`testcase.Case.Assert()` is called.
 
 ### Changed
 - Return the exitreason from GenCaller, not the whole exitmsg.